Detailed API

HarmonicPowerFlow — Module

A simple and fast module to solve the harmonic power flow problem for given distribution grids.

HarmonicPowerFlow.THD — Method

THD(u)

Calculate the Total Harmonic Distortion at all nodes.

THD is a measure for the amount of voltage distortion present at a node. This function calculates two alternative definitions - THDF: relative to fundamental, more commonly used and - THDR: RMS, relative to all frequencies.

HarmonicPowerFlow.admittance_matrices — Method

admittance_matrices(nodes, lines, harmonics)

Build the nodal admittance matrices (admittance laplacian) for all harmonics. Admittance scales with frequency (Xh = Xf * h).

Returns: dictionary of DataFrames LY[h]

HarmonicPowerFlow.barplot_THD — Method

Barplot of harmonic voltage magnitude at one bus

HarmonicPowerFlow.create_nodes_manually — Method

createnodesmanually()

Manually create a nodes DataFrame.

Note that linear nodes are added first, then nonlinear ones. Within the linear ones, first add PV, then PQ nodes. First node contains the slack bus.

HarmonicPowerFlow.current_injections — Method

calculate the harmonic current injections at one node

HarmonicPowerFlow.fund_state_vec — Method

fund_state_vec(u)

Take the voltage DataFrame and return a complex vector.

Note: sorting of values. For the fundamental state vector voltage phase comes first, then magnitude.

HarmonicPowerFlow.heatmap_NE — Method

Heatmap indicating magnitude of Norton Equivalent parameters

HarmonicPowerFlow.hpf — Method

hpf(net, settings)

Solve the harmonic power flow problem for a given power grid and settings.

HarmonicPowerFlow.import_Norton_Equivalents — Function

Import Norton Equivalent parameters for all nonlinear devices in "nodes" from folder

HarmonicPowerFlow.init_power_grid — Method

init_power_grid(nodes, lines, settings)

Create a PowerGrid struct from nodes and lines DataFrames and converts all quantities to the p.u. system. Indexing (c, m, n) could be done better by being based on number of nodes of certain type. It would make sense to also add LY to net struct.

HarmonicPowerFlow.limits — Method

EN50160(u, nodeID)

Check if harmonic limits according to EN 50160 and EN 61000 are met at node.